# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Comment System API",
    "version": "1.0.0",
    "description": "A comprehensive comment system API with authentication, comments, likes, and real-time updates via WebSocket",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token. Format: Bearer {token}"
      }
    },
    "schemas": {
...
      "Like": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Like unique identifier",
            "example": "clm1234567890abc"
          },
          "userId": {
            "type": "string",
            "description": "User ID who created like",
            "example": "clm1234567890abc"
          },
          "commentId": {
            "type": "string",
            "description": "Comment ID (if liking a comment)",
            "example": "clm1234567890abc"
          },
          "replyId": {
            "type": "string",
            "description": "Reply ID (if liking a reply)",
            "example": "clm1234567890abc"
          },
          "type": {
            "type": "string",
            "enum": [
              "like",
              "dislike"
            ],
            "description": "Type of like",
            "example": "like"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Like creation date",
            "example": "2024-01-01T00:00:00.000Z"
          }
        }
      },
      "LikeCounts": {
        "type": "object",
        "properties": {
          "likesCount": {
            "type": "number",
            "description": "Number of likes",
            "example": 5
          },
          "dislikesCount": {
            "type": "number",
            "description": "Number of dislikes",
            "example": 2
          },
          "userLikeStatus": {
            "type": "object",
            "properties": {
              "hasLiked": {
                "type": "boolean",
                "description": "Whether the current user has liked",
                "example": true
              },
              "hasDisliked": {
                "type": "boolean",
                "description": "Whether the current user has disliked",
                "example": false
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Error code",
            "example": "ERROR_CODE"
          },
          "errors": {
            "type": "array",
            "description": "Validation errors (if applicable)",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field name",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "description": "Error message",
                  "example": "Invalid email format"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/likes/comment/{commentId}": {
      "post": {
        "summary": "Like a comment",
        "description": "Add a like to a comment",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comment ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Comment liked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment liked successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "like": {
                          "$ref": "#/components/schemas/Like"
                        },
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove like/dislike from a comment",
        "description": "Remove a like or dislike from a comment",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Like/dislike removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Like/dislike removed from comment successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Like not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/likes/comment/{commentId}/dislike": {
      "post": {
        "summary": "Dislike a comment",
        "description": "Add a dislike to a comment",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comment ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Comment disliked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment disliked successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "like": {
                          "$ref": "#/components/schemas/Like"
                        },
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/likes/comment/{commentId}/status": {
      "get": {
        "summary": "Get like status for current user",
        "description": "Check if the current user has liked or disliked a comment",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Like status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment like status retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "hasLiked": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasDisliked": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/likes/reply/{replyId}": {
      "post": {
        "summary": "Like a reply",
        "description": "Add a like to a reply",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "replyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reply ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Reply liked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Reply liked successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "like": {
                          "$ref": "#/components/schemas/Like"
                        },
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reply not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove like/dislike from a reply",
        "description": "Remove a like or dislike from a reply",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "replyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reply ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Like/dislike removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Like/dislike removed from reply successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Like not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/likes/reply/{replyId}/dislike": {
      "post": {
        "summary": "Dislike a reply",
        "description": "Add a dislike to a reply",
        "tags": [
          "Likes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "replyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Reply ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Reply disliked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Reply disliked successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "like": {
                          "$ref": "#/components/schemas/Like"
                        },
                        "counts": {
                          "$ref": "#/components/schemas/LikeCounts"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reply not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}